+++ easyink-wecom/src/main/java/com/easyink/wecom/domain/entity/form/WeFormUseRecord.java
@@ -28,7 +28,7 @@ public class WeFormUseRecord {
     /**
      * 表单id
      */
-    private Integer formId;
+    private Long formId;
     /**
      * 员工id
      */
